Refactor test structure for PlotContexts in kandy-api and add LayerPlotContext test
